@@ -1,58 +1,101 @@
 {
-  "name": "klirtechchallenge",
-  "version": "0.0.0",
+  "name": "klir-tech-challenge-web-front",
+  "version": "1.0.0",
+  "private": true,
   "scripts": {
     "ng": "ng",
-    "start": "ng serve",
-    "build": "ng build",
-    "build:ssr": "ng run KlirTechChallenge:server:dev",
-    "test": "ng test",
-    "lint": "ng lint",
-    "e2e": "ng e2e"
+    "build": "npm run write:env -s && ng build --prod",
+    "start": "npm run write:env -s && ng serve",
+    "serve:sw": "npm run build -s && npx http-server ./dist -p 4200",
+    "lint": "ng lint && stylelint \"src/**/*.scss\" --syntax scss && htmlhint \"src\" --config .htmlhintrc",
+    "test": "npm run write:env -s && ng test",
+    "test:ci": "npm run write:env -s && npm run lint -s && ng test --configuration=ci",
+    "e2e": "npm run write:env -s && ng e2e",
+    "translations:extract": "ngx-translate-extract --input ./src --output ./src/translations/template.json --format=json --clean --sort",
+    "docs": "hads ./docs -o",
+    "write:env": "ngx-scripts env npm_package_version",
+    "prettier": "prettier --write \"./{src,e2e}/**/*.{ts,js,html,scss}\"",
+    "prettier:check": "prettier --list-different \"./{src,e2e}/**/*.{ts,js,html,scss}\"",
+    "postinstall": "npm run prettier -s",
+    "generate": "ng generate"
   },
-  "private": true,
   "dependencies": {
-    "@angular/animations": "8.2.12",
-    "@angular/common": "8.2.12",
-    "@angular/compiler": "8.2.12",
-    "@angular/core": "8.2.12",
-    "@angular/forms": "8.2.12",
-    "@angular/platform-browser": "8.2.12",
-    "@angular/platform-browser-dynamic": "8.2.12",
-    "@angular/platform-server": "8.2.12",
-    "@angular/router": "8.2.12",
-    "@nguniversal/module-map-ngfactory-loader": "8.1.1",
-    "aspnet-prerendering": "^3.0.1",
+    "@angular/animations": "^11.0.8",
+    "@angular/common": "^11.0.8",
+    "@angular/compiler": "^11.0.8",
+    "@angular/core": "^11.0.8",
+    "@angular/forms": "^11.0.8",
+    "@angular/localize": "^11.0.8",
+    "@angular/material": "^12.0.4",
+    "@angular/platform-browser": "^11.0.8",
+    "@angular/platform-browser-dynamic": "^11.0.8",
+    "@angular/router": "^11.0.8",
+    "@angular/service-worker": "^11.0.8",
+    "@fortawesome/fontawesome-free": "^5.15.1",
+    "@ng-bootstrap/ng-bootstrap": "^8.0.0",
+    "@ngx-translate/core": "^13.0.0",
+    "angular-datatables": "^12.0.0",
     "bootstrap": "^4.6.0",
-    "core-js": "^3.8.3",
-    "jquery": "^3.5.1",
-    "oidc-client": "^1.11.3",
-    "popper.js": "^1.16.0",
+    "datatables.net": "^1.10.25",
+    "datatables.net-dt": "^1.10.25",
     "rxjs": "^6.6.3",
-    "zone.js": "0.9.1"
+    "tslib": "^2.0.1",
+    "zone.js": "~0.10.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.803.26",
-    "@angular/cli": "^8.3.29",
-    "@angular/compiler-cli": "^8.2.14",
-    "@angular/language-service": "^8.2.14",
-    "@types/jasmine": "~3.4.4",
-    "@types/jasminewd2": "~2.0.8",
-    "@types/node": "~12.11.6",
-    "codelyzer": "^5.2.2",
-    "jasmine-core": "~3.5.0",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "^5.2.3",
+    "@angular-devkit/build-angular": "^12.0.4",
+    "@angular/cli": "~11.0.6",
+    "@angular/compiler-cli": "^11.0.8",
+    "@angular/language-service": "^11.0.8",
+    "@biesbjerg/ngx-translate-extract": "^7.0.3",
+    "@biesbjerg/ngx-translate-extract-marker": "^1.0.0",
+    "@ngneat/until-destroy": "^8.0.3",
+    "@ngx-rocket/scripts": "^5.0.0",
+    "@types/datatables.net": "^1.10.19",
+    "@types/jasmine": "^3.6.2",
+    "@types/jasminewd2": "^2.0.8",
+    "@types/jquery": "^3.5.5",
+    "@types/node": "^12.19.12",
+    "codelyzer": "^6.0.1",
+    "hads": "^3.0.0",
+    "htmlhint": "^0.9.9",
+    "https-proxy-agent": "^5.0.0",
+    "husky": "^4.3.6",
+    "jasmine-core": "~3.6.0",
+    "jasmine-spec-reporter": "~6.0.0",
+    "karma": "^6.3.4",
     "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage-istanbul-reporter": "~2.1.0",
-    "karma-jasmine": "~2.0.1",
+    "karma-coverage-istanbul-reporter": "~3.0.3",
+    "karma-jasmine": "~4.0.1",
     "karma-jasmine-html-reporter": "^1.5.4",
-    "typescript": "3.5.3"
+    "karma-junit-reporter": "^2.0.1",
+    "prettier": "^2.2.1",
+    "pretty-quick": "^3.1.0",
+    "protractor": "~7.0.0",
+    "stylelint": "^9.3.0",
+    "stylelint-config-prettier": "^7.0.0",
+    "stylelint-config-recommended-scss": "^3.3.0",
+    "stylelint-config-standard": "^18.3.0",
+    "stylelint-scss": "~3.18.0",
+    "ts-node": "^9.1.1",
+    "tslint": "~6.1.0",
+    "tslint-config-prettier": "^1.18.0",
+    "typescript": "~4.0.5"
   },
-  "optionalDependencies": {
-    "node-sass": "^5.0.0",
-    "protractor": "~5.4.2",
-    "ts-node": "~8.4.1",
-    "tslint": "~5.20.0"
+  "prettier": {
+    "singleQuote": true,
+    "overrides": [
+      {
+        "files": "*.scss",
+        "options": {
+          "singleQuote": false
+        }
+      }
+    ]
+  },
+  "husky": {
+    "hooks": {
+      "pre-commit": "pretty-quick --staged"
+    }
   }
 }
