@@ -1,21 +1,19 @@
+/*
+ * Entry point of the application.
+ * Only platform bootstrapping code should be here.
+ * For app-specific initialization, use `app/app.component.ts`.
+ */
+
 import { enableProdMode } from '@angular/core';
 import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
 
-import { AppModule } from './app/app.module';
-import { environment } from './environments/environment';
-
-export function getBaseUrl() {
-  return document.getElementsByTagName('base')[0].href;
-}
-
-const providers = [
-  { provide: 'BASE_URL', useFactory: getBaseUrl, deps: [] },
-  { provide: 'BASE_API_URL', useValue: environment.apiUrl }
-];
+import { AppModule } from '@app/app.module';
+import { environment } from '@env/environment';
 
 if (environment.production) {
   enableProdMode();
 }
 
-platformBrowserDynamic(providers).bootstrapModule(AppModule)
-  .catch(err => console.log(err));
+platformBrowserDynamic()
+  .bootstrapModule(AppModule)
+  .catch((err) => console.error(err));
